requirements.txt
africastalking==1.2.5
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
contextlib2==21.6.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1
django-debug-toolbar==3.6.0
django-filter==22.1
django-generic-plus==2.4.0
django-grappelli==3.0.3
django-import-export==2.8.0
django-jazzmin==2.5.0
django-redis==5.2.0
django-serverside-datatable==2.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
drf-generators==0.5.0
drf-spectacular==0.23.1
drf_generator==0.0.7
et-xmlfile==1.1.0
gunicorn==20.1.0
hiredis==2.0.0
idna==3.4
inflection==0.5.1
jsonschema==4.14.0
Markdown==3.4.1
MarkupPy==1.14
mysql-connector-python==8.0.30
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
protobuf==3.20.1
psycopg2==2.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.21.0
python-monkey-business==1.0.0
python-xmp-toolkit==2.0.1
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.2
schema==0.7.5
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0